Update times to be float64s for timing precision better than 10 seconds
Note that I left the barycentric correction as a float32. This is in line with the data types of the official TESS short cadence light curves, which also have tstart and tstop as 64 bit but barycorr as 32. This is because the 8 digits of barycentric precision is enough to get to the 11th decimal place, since they always start as 0.000XXX.